fix: add fallback for loading favicon image
