+++ apps/chat/src/pages/api/themes/favicon.ts
@@ -2,19 +2,13 @@ import { NextApiRequest, NextApiResponse } from 'next';
 
 import { isAbsoluteUrl } from '@/src/utils/app/file';
 import { getImageUrl } from '@/src/utils/app/themes';
-import { logger } from '@/src/utils/server/logger';
+import { withThemesConfig } from '@/src/utils/server/themes-config';
 
-import { HTTPMethod } from '@/src/types/http';
 import { ThemesConfig } from '@/src/types/themes';
-
-import { errorsMessages } from '@/src/constants/errors';
 
 import fetch from 'node-fetch';
 
 const FAVICON_NAMES = ['chat-favicon', 'favicon'] as const;
-
-let cachedTheme: ThemesConfig | undefined = undefined;
-let cachedThemeExpiration: number | undefined;
 
 const fetchFavicon = async (
   res: NextApiResponse,
@@ -45,52 +39,7 @@ const fetchFavicon = async (
 };
 
 const handler = async (_req: NextApiRequest, res: NextApiResponse) => {
-  try {
-    if (!process.env.THEMES_CONFIG_HOST) {
-      return res.status(500).send(errorsMessages.customThemesConfigNotProvided);
-    }
-
-    if (
-      cachedThemeExpiration &&
-      cachedTheme &&
-      cachedThemeExpiration > Date.now()
-    ) {
-      return fetchFavicon(res, cachedTheme);
-    }
-
-    const controller = new AbortController();
-    const response = await fetch(
-      `${process.env.THEMES_CONFIG_HOST}/config.json`,
-      {
-        method: HTTPMethod.GET,
-        headers: {
-          'Content-Type': 'application/json',
-          'Cache-Control': 'public, max-age=604800',
-        },
-        signal: controller.signal,
-      },
-    );
-
-    if (!response.ok) {
-      logger.error(
-        `Received error when fetching config file: ${response.status} ${
-          response.statusText
-        } ${await response.text()}`,
-      );
-      return res.status(500).send(errorsMessages.generalServer);
-    }
-
-    const json = (await response.json()) as ThemesConfig;
-
-    const dayInMs = 86400000;
-    cachedThemeExpiration = Date.now() + dayInMs;
-    cachedTheme = json;
-
-    return fetchFavicon(res, cachedTheme);
-  } catch (e) {
-    logger.error(e);
-    return res.status(500).send(errorsMessages.generalServer);
-  }
+  return withThemesConfig(res, (theme) => fetchFavicon(res, theme));
 };
 
 export default handler;
