+++ apps/chat/src/pages/api/themes/image/[name].ts
@@ -2,17 +2,11 @@ import { NextApiRequest, NextApiResponse } from 'next';
 
 import { isAbsoluteUrl } from '@/src/utils/app/file';
 import { getImageUrl } from '@/src/utils/app/themes';
-import { logger } from '@/src/utils/server/logger';
+import { withThemesConfig } from '@/src/utils/server/themes-config';
 
-import { HTTPMethod } from '@/src/types/http';
 import { ThemesConfig } from '@/src/types/themes';
 
-import { errorsMessages } from '@/src/constants/errors';
-
 import fetch from 'node-fetch';
-
-let cachedTheme: ThemesConfig | undefined = undefined;
-let cachedThemeExpiration: number | undefined;
 
 const getImage = async (
   _req: NextApiRequest,
@@ -47,60 +41,13 @@ const getImage = async (
 };
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
-  try {
-    if (!process.env.THEMES_CONFIG_HOST) {
-      return res.status(500).send(errorsMessages.customThemesConfigNotProvided);
-    }
+  const name = req.query.name;
 
-    const name = req.query.name;
-
-    if (!name || Array.isArray(name)) {
-      return res
-        .status(500)
-        .send('Name parameter not provided for theme image');
-    }
-
-    if (
-      cachedThemeExpiration &&
-      cachedTheme &&
-      cachedThemeExpiration > Date.now()
-    ) {
-      return getImage(req, res, cachedTheme, name);
-    }
-
-    const controller = new AbortController();
-    const response = await fetch(
-      `${process.env.THEMES_CONFIG_HOST}/config.json`,
-      {
-        method: HTTPMethod.GET,
-        headers: {
-          'Content-Type': 'application/json',
-          'Cache-Control': 'public, max-age=604800',
-        },
-        signal: controller.signal,
-      },
-    );
-
-    if (!response.ok) {
-      logger.error(
-        `Received error when fetching config file: ${response.status} ${
-          response.statusText
-        } ${await response.text()}`,
-      );
-      return res.status(500).send(errorsMessages.generalServer);
-    }
-
-    const json = (await response.json()) as ThemesConfig;
-
-    const dayInMs = 86400000;
-    cachedThemeExpiration = Date.now() + dayInMs;
-    cachedTheme = json;
-
-    return getImage(req, res, cachedTheme, name);
-  } catch (e) {
-    logger.error(e);
-    return res.status(500).send(errorsMessages.generalServer);
+  if (!name || Array.isArray(name)) {
+    return res.status(500).send('Name parameter not provided for theme image');
   }
+
+  return withThemesConfig(res, (theme) => getImage(req, res, theme, name));
 };
 
 export default handler;
